requirements.txt
streamlit
openai
requests
pillow